Fix: modulos are 15-bit signed, the minterms are also in regular BLTCON0.

+++ Machines/Amiga/Blitter.hpp
@@ -44,7 +44,7 @@ class Blitter: public DMADevice<4> {
 		uint8_t minterms_ = 0;
 		int width_ = 0, height_ = 0;
 		uint32_t a_ = 0, b_ = 0;
-		uint16_t modulos_[4]{};
+		uint32_t modulos_[4]{};
 
 		int shifts_[2]{};
 		bool line_mode_ = false;
